Make CI strictly green: fix all mypy errors, drop continue-on-error
CI's first runs were green per the job-status (X step had continue-on-error)
but the mypy step was silently producing 50+ errors. This commit fixes them
all and removes the bypass.

Configuration
-------------
- pyproject.toml [tool.mypy]: relax `disallow_any_unimported = false` —
  gymnasium and torch ship without stubs, so the strict-mode default would
  require # type: ignore on every Env / nn.Module subclass. Cost is bounded
  Any-propagation across those library boundaries; our own code stays strict.
- pyproject.toml [[tool.mypy.overrides]]: add `reflexive_options.third_party.*`
  with `ignore_errors = true`. Vendored ATLAS / RAT keep upstream typing per
  CLAUDE.md vendoring discipline. Also add gymnasium / torch / pyarrow / ripser
  to ignore_missing_imports.
- .github/workflows/ci.yml: drop `continue-on-error: true` from the mypy step
  so future regressions actually fail CI.

Code fixes (real)
-----------------
- theory/bifurcation.py, simulator/coupling.py: cast eigvals to complex128
  (numpy returns float64 for symmetric input).
- experiments/_common.py: narrow `is_dataclass(config)` with `not isinstance
  (config, type)` so `asdict` accepts the instance.
- baselines/lsv.py: branch on `np.ndim(result) == 0` instead of np.isscalar
  (cleaner narrowing for the float-or-array return).
- simulator/gamma_aggregator.py: declare `_oi_calls` / `_oi_puts` at the class
  level so the second-branch `.astype` doesn't widen the inferred 2D shape.
- Several call sites: explicit `.astype(np.float64)` on `np.linspace` /
  `np.geomspace` outputs (numpy 2 stubs return floating[Any]).
- experiments/{phase_diagram,reflexive_transfer,synthetic_replication,
  bifurcation_scan}.py: replace bare `np.ndarray` with `NDArray[np.float64]`.
- rl/env.py, experiments/reflexive_transfer.py: subclassing of `gym.Env` and
  `nn.Module` annotated `# type: ignore[misc]`.
- surface/io.py, rl/{env,actions}.py: drop now-unused type-ignore comments.

Verification
------------
- ruff 0.15.12: clean across src + tests.
- mypy 1.20.2: `Success: no issues found in 64 source files`.
- pytest: 116/116 passing in 112s.

+++ src/reflexive_options/rl/curriculum.py
@@ -143,10 +143,9 @@ def _default_surface_grid() -> SurfaceGrid:
     """7 maturities × 11 strikes, matching `paper/pre_registration.md` §4."""
     return SurfaceGrid(
         log_moneyness=np.linspace(-0.20, 0.20, 11, dtype=np.float64),
-        maturities=np.array(
-            [7, 14, 30, 60, 90, 180, 365], dtype=np.float64
-        )
-        / 365.0,
+        maturities=(np.array([7, 14, 30, 60, 90, 180, 365], dtype=np.float64) / 365.0).astype(
+            np.float64
+        ),
     )
